refactor: split coaching module into single-responsibility files
- Split CoachingExceptionHandler into CoachNotFoundExceptionHandler and
  ValidationExceptionHandler
- Extract toResponse() extension functions from response DTOs into
  dedicated files under dtos/extensions
- Split CoachingEventListener into per-event listener classes
