+++ src/main/kotlin/com/nickdferrara/fitify/coaching/internal/dtos/response/CertificationResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.coaching.internal.dtos.response
 
-import com.nickdferrara.fitify.coaching.internal.entities.CoachCertification
 import java.time.LocalDate
 import java.util.UUID
 
@@ -9,11 +8,4 @@ internal data class CertificationResponse(
     val name: String,
     val issuer: String,
     val validUntil: LocalDate?,
-)
-
-internal fun CoachCertification.toResponse() = CertificationResponse(
-    id = id!!,
-    name = name,
-    issuer = issuer,
-    validUntil = validUntil,
 )
