+++ src/main/kotlin/com/nickdferrara/fitify/coaching/internal/dtos/response/CoachResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.coaching.internal.dtos.response
 
-import com.nickdferrara.fitify.coaching.internal.entities.Coach
 import java.time.Instant
 import java.util.UUID
 
@@ -14,16 +13,4 @@ internal data class CoachResponse(
     val createdAt: Instant,
     val certifications: List<CertificationResponse>,
     val locationIds: List<UUID>,
-)
-
-internal fun Coach.toResponse() = CoachResponse(
-    id = id!!,
-    name = name,
-    bio = bio,
-    photoUrl = photoUrl,
-    specializations = specializations,
-    active = active,
-    createdAt = createdAt!!,
-    certifications = certifications.map { it.toResponse() },
-    locationIds = locations.map { it.locationId },
 )
